Make sure the crop person exists before deleting it.
IBP-2784

+++ src/main/java/org/generationcp/ibpworkbench/service/ProgramService.java
@@ -180,8 +180,10 @@ public class ProgramService {
 		// Get the users with no association to any programs in a crop.
 		final List<WorkbenchUser> usersWithoutAssociatedPrograms = this.userService.getUsersWithoutAssociatedPrograms(project.getCropType());
 		for (final WorkbenchUser workbenchUser : usersWithoutAssociatedPrograms) {
-			final CropPerson cropPerson = new CropPerson(project.getCropType(), workbenchUser.getPerson());
-			this.userService.removeCropPerson(cropPerson);
+			final CropPerson cropPerson = this.userService.getCropPerson(project.getCropType().getCropName(), workbenchUser.getPerson().getId());
+			if (cropPerson != null) {
+				this.userService.removeCropPerson(cropPerson);
+			}
 		}
 	}
